+++ ui/src/main/res/navigation/navigation_main.xml
@@ -6,6 +6,6 @@
 
     <fragment
         android:id="@+id/mainFragment"
-        android:name="com.nemo.githubrepositories.main.MainListFragment"
+        android:name="com.nemo.githubrepositories.main.list.MainListFragment"
         android:label="MainFragment" />
 </navigation>
